Separate build script and simplified PrgEnv-fluidity module
Do the build with a bash script to configure and a PBS script to
compile.  Using RPATH/RUNPATH in the libraries and executables means
the the PrgEnv-fluidity module is simplified.  The Zoltan library from
the cray-trilinos module that matches the cray-petsc module is used
(the other Trilinos packages are not loaded).

Some test scripts are available but have not been tested.